[Ref] More small changes to get rid of warnings in 64-bit compile
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@2598 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ common/serialization_utils.cpp
@@ -172,7 +172,7 @@ void ReadAdaptive1248(std::istream& iStrm, uint64& val)
 void WriteItemString(std::ostream& oStrm, const char* const pStr, const size_t nSize)
 //--------------------------------------------------------------------------------
 {
-	uint32 id = std::min<size_t>(nSize, (uint32_max >> 4)) << 4;
+	uint32 id = (uint32)std::min<size_t>(nSize, (uint32_max >> 4)) << 4;
 	id |= 12; // 12 == 1100b
 	Binarywrite<uint32>(oStrm, id);
 	id >>= 4;
@@ -206,7 +206,7 @@ void ReadItemString(std::istream& iStrm, std::string& str, const DataSize)
 	for(size_t i = 0; i < str.size(); i++)
 		iStrm.read(&str[i], 1);
 
-	id = (id >> 4) - str.size();
+	id = (id >> 4) - (uint32)str.size();
 	if(id > 0)
 		iStrm.ignore(id);
 }
@@ -585,7 +585,7 @@ void Ssb::OnWroteItem(const char* pId, const size_t nIdSize, const Postype& posB
 	const Offtype nRawEntrySize = m_pOstrm->tellp() - posBeforeWrite;
 
 	if (nRawEntrySize > std::numeric_limits<DataSize>::max())
-		{ AddWriteNote(SNW_INSUFFICIENT_DATASIZETYPE); return; }	
+		{ AddWriteNote(SNW_INSUFFICIENT_DATASIZETYPE); return; }
 
 	if(GetFlag(RwfRMapHasSize) && nRawEntrySize > (std::numeric_limits<DataSize>::max() >> 2))
 		{ AddWriteNote(SNW_DATASIZETYPE_OVERFLOW); return; }
